Move location of dependencies dec
