+++ scvi/data/_preprocessing.py
@@ -9,9 +9,8 @@ import pandas as pd
 import torch
 from scipy.sparse import issparse
 
-from scvi._decorators import dependencies
 from scvi.model._utils import parse_device_args
-from scvi.utils import track
+from scvi.utils import dependencies, track
 from scvi.utils._docstrings import devices_dsp
 
 from ._utils import _check_nonnegative_integers
