+++ scvi/utils/__init__.py
@@ -1,6 +1,6 @@
 from ._attrdict import attrdict
 from ._decorators import unsupported_if_adata_minified
-from ._dependencies import error_on_missing_dependencies
+from ._dependencies import dependencies, error_on_missing_dependencies
 from ._docstrings import de_dsp, setup_anndata_dsp
 from ._exceptions import InvalidParameterError
 from ._jax import device_selecting_PRNGKey
@@ -15,4 +15,5 @@ __all__ = [
     "unsupported_if_adata_minified",
     "InvalidParameterError",
     "error_on_missing_dependencies",
+    "dependencies",
 ]
